BLl 2nd

+++ WindowSeller/WindowSeller.Domain/SubElement.cs
@@ -8,7 +8,7 @@ namespace WindowSeller.Domain
 {
     public partial class SubElement
     {
-        public int SubElementId { get; set; }
+        public long SubElementId { get; set; }
         public int Element { get; set; }
         public decimal Width { get; set; }
         public decimal Height { get; set; }
