+++ WindowSeller/WindowSellerWASM/Shared/Persistance/IWindowRepository.cs
@@ -9,6 +9,6 @@ namespace WindowSellerWASM.Shared.Persistance
 {
     public interface IWindowRepository: IGenericRepository<Window>
     {
-        Task<List<Window>> GetByOrderIdAsync(int orderId);
+        Task<List<Window>> GetByOrderIdAsync(long orderId);
     }
 }
